Added Alarm Class ====================
Added the Alarm Class which is,currently, a thin wrapper
around an AVAudioPlayer.

Verified that it is constructed without issue in a viewDidLoad
method (probably need to alter).

Additionally, added a resource to be red (test.wav). Lydia
was sleeping when I downloaded it, and my computer was on mute.
No clue what the sound is.